+++ R/api-make.R
@@ -163,7 +163,9 @@ make <- function(
   memory_strategy = "speed",
   layout = NULL,
   lock_envir = TRUE,
-  history = TRUE
+  history = TRUE,
+  recover = FALSE,
+  recoverable = TRUE
 ) {
   log_msg(
     "begin make()",
@@ -219,7 +221,9 @@ make <- function(
       memory_strategy = memory_strategy,
       layout = layout,
       lock_envir = lock_envir,
-      history = history
+      history = history,
+      recover = recover,
+      recoverable = recoverable
     )
   }
   assert_config_not_plan(config)