+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## New features
 
+- Add automated data recovery (#945). This is still experimental and disabled by default. Requires `make(recover = TRUE)`.
+- Add new functions `recoverable()` and `r_recoverable()` to show targets that are outdated but recoverable via `make(recover = TRUE)`.
 - Track the history and provenance of targets, viewable with `drake_history()`. Powered by [`txtq`](https://github.com/wlandau/txtq) (#918, #920).
 - Add a new `no_deps()` function, similar to `ignore()`. `no_deps()` suppresses dependency detection but still tracks changes to the literal code ([#910](https://github.com/ropensci/drake/issues/910)).
 - Add a new "autoclean" memory strategy (#917).